only generate template when needed
